exports 'devkit' submodule, update version release
