+++ lib/devkit/devkit.models.ts
@@ -1,0 +1,6 @@
+import type { HttpServiceStatus, JsonValue } from '../public_api.js';
+
+export type FetchResponse = {
+	readonly statusCode: HttpServiceStatus;
+	readonly json: JsonValue;
+};
