+++ lib/devkit_api.ts
@@ -1,0 +1,5 @@
+/**
+ * Devkit API - used for development purposes in
+ */
+export type { getFakeBlob, getFetchBlobMock, getFetchJsonMock } from './devkit/test.utils.js';
+export { replaceSdkVersionPlaceholder } from './utils/sdk-version.utils.js';
