+++ lib/sdk-info.ts
@@ -1,5 +1,5 @@
 export const sdkInfo = {
 	host: 'npmjs.com',
-	version: '10.10.0',
+	version: '{{version}}',
 	name: '@kontent-ai/core-sdk',
 };
