+++ scripts/clean.ts
@@ -1,4 +1,4 @@
-import { deleteFolderRecursive } from './script.utils.js';
+import { deleteFolderRecursive } from "../lib/devkit/script.utils.js";
 
 for (const path of ['dist']) {
 	deleteFolderRecursive(path);
