+++ scripts/update-version.ts
@@ -1,9 +1,4 @@
+import { replaceSdkVersionPlaceholder } from '../lib/devkit_api.js';
 import packageJson from '../package.json' with { type: 'json' };
-import { createVersionFile } from './script.utils.js';
 
-createVersionFile({
-	filePath: './lib/sdk.ts',
-	propertyName: 'sdkInfo',
-	packageName: packageJson.name,
-	packageVersion: packageJson.version,
-});
+replaceSdkVersionPlaceholder('./dist/sdk-info.js', packageJson.version);
