+++ tests/headers/headers.spec.ts
@@ -1,8 +1,8 @@
 import { afterAll, describe, expect, it, vi } from 'vitest';
+import { getFetchJsonMock } from '../../lib/devkit/test.utils.js';
 import { defaultHttpService } from '../../lib/public_api.js';
-import { sdkInfo } from '../../lib/sdk.generated.js';
+import { sdkInfo } from '../../lib/sdk-info.js';
 import { getSdkIdHeader } from '../../lib/utils/header.utils.js';
-import { getFetchJsonMock } from '../_utils/test.utils.js';
 
 const sdkIdHeader = getSdkIdHeader(sdkInfo);
